MetadataHandler was missing a flush() after write.
Note: This has the potential to slow down the log server since it
introduces flushing for each metadata entry that gets written, however
I don't think we have any actual choice in the matter: we have to
flush.  If performance issues arise we might have to add cleverness to
this.