Better argument validation for ZkTool.
Make ZkTool verify that the combination of arguments is valid and
specifies one and only one action for ZkTool to perform. Print an error
message and exit if not. Also, listing coordinates is now done with
"ZkTool --list" instead of requiring an unused coordinate option with
"ZkTool --operation LIST --coordinate 1.some.bogus.coordinate".