Fail when we are not able to create slot directory
When we get a new slot for a logfile we have to ensure that the
directory of the slotfile path exists, if it doesnt exist we try to
create the directories. We already did all of this, but we did not check
if the creation of the directories was successful. This commit fails we
are unable to create all of the directories of a slotfile.