Log slot numbering based on first log event timestamp.
Log slots were postfixed with a creation timestamp, which should
always be a few milliseconds off the timestamp of the first log
event. Instead of this postfix with the first log timestamp.
This also fixes a test case where we used System.currentMillis
instead of a TimeProvider to set the previous name scheme.