Flags through properties file gets no special treatment.
Handlig of argument, wether it's from the command line or via
a properties file, should be handled in the same way. If you
expect a certain result from the command line, you should be
expecting the same result of you simply put that argument in
a properties file.

- Can now do "help" and "version". There are no big issues in
allowing the use of those flags through that file.
- Will now crash if an argument is given for which its Flag
counterpart does not exist.