Proof-of-concept Consul backend
This is a working Consul backend for Cloudname but it hasn't been tested
a lot. In short it creates a session which is kept alive by regular polls
to the Consul Agent. The session is then used to create ephmeral KV entries.

Permanent leases are just regular KV entries without an associated session.

Note that this implementation isn't tested in a real life production
environment and that the defaults are most likely a bit too aggressive for
to be used as is. In particular the polling when a watch gets 404 is very
aggressive and the lock-delay parameter is set to 0.

There's no integration with the service catalog in Consul either since that
is handled on a higher level in Cloudname. This might require a slight change
in the core API with some sort of hooks or custom service implementation on
top of Cloudname. A custom service implementation isn't ideal.

The backend tests are also updated a bit with more generous backend propagation
timeouts throughout.