Simplify core interface
Merge the permanent/temporary methods into a single coherent set. This makes
the interface a bit more flexible since you can listen on collections of
leases of both kinds. Since the backend doesn't have to implement naming
of temporary leases and store all of the leases in the same name space the
various implementations are a lot simpler as well.

In addition - javadoc reporting pluging added. Some of the issues reported
as warnings was fixed but in the end linting was turned off. It should be
turned on later on when there's been a general cleanup. Run with
mvn javadoc:aggregate to generate javadoc locally.